refactor: make ajusts to implements the screen flow
