+++ src/cli/data/creator-params.ts
@@ -1,3 +1,7 @@
 export interface ComponentCliParams {
   components: string[]
 }
+
+export interface ScreenCliParams {
+  screens: string[]
+}
