+++ src/common/functions/validation.ts
@@ -7,3 +7,11 @@ export const validateComponentName = (userInput:any) => {
 
   return true
 }
+
+export const formatComponentsNames = (componentsNames: string[]): string[] => {
+  return componentsNames.map( componentName => {
+    const splitedName = componentName.split('-');
+    const compositeName = splitedName.reduce( (response, name) => response + name.substring(0,1).toUpperCase() + name.substring(1), '');
+    return compositeName
+  })
+}
